Added Appium Server Initialization
