+++ src/main/java/pages/LoginPage.java
@@ -1,13 +1,13 @@
 package pages;
 
 
-import BaseClass.BaseTest;
+import BaseClass.TestBotBase;
 import io.appium.java_client.MobileElement;
 import io.appium.java_client.pagefactory.AndroidFindBy;
 import io.appium.java_client.pagefactory.iOSXCUITFindBy;
 import utils.TestUtils;
 
-public class LoginPage extends BaseTest {
+public class LoginPage extends TestBotBase {
 	TestUtils utils = new TestUtils();
 	@AndroidFindBy (accessibility = "test-Username") 
 	@iOSXCUITFindBy (id = "test-Username")
