--- src/test/resources/testng.xml
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE suite>
-<suite name="Suite" parallel="tests">
-    <test name="Pixel_2">
-        <listeners>
-            <listener class-name="com.qa.listeners.TestListener" />
-        </listeners>
-        <parameter name="emulator" value="true" />
-        <parameter name="platformName" value="Android" />
-        <parameter name="udid" value="58H3Y18720017229" />
-        <parameter name="deviceName" value="Honor Play" />
-        <parameter name="systemPort" value="10000" />
-        <parameter name="chromeDriverPort" value="11000" />
-        <classes>
-            <class name="com.qa.tests.LoginTests"/>
-            <class name="com.qa.tests.ProductTests"/>
-        </classes>
-    </test>
-
-</suite> <!-- Suite -->
